[#3] Make project MVP: fixed podcasts stat on main and podcast's list pages
