+++ src/schemas.py
@@ -1,7 +1,9 @@
 from datetime import datetime
 from typing import Optional
 
-from pydantic import BaseModel, Field
+from pydantic import BaseModel, Field, computed_field
+
+from src.constants import format_file_size
 
 __all__ = (
     "SystemInfo",
@@ -28,3 +30,38 @@ class ErrorResponse(BaseModel):
 
     error: str
     detail: Optional[str] = None
+
+
+class RecentActivity(BaseModel):
+    """Recent activity text and optional time string for display."""
+
+    text: str
+    time: str | None = None
+
+
+class AppStatistics(BaseModel):
+    """Application-wide statistics for dashboard (index)."""
+
+    total_episodes: int = 0
+    total_podcasts: int = 0
+    total_duration: int = 0
+    total_size: int = 0
+    downloading_count: int = 0
+    last_published_at: datetime | None = None
+    last_created_at: datetime | None = None
+    recent_activity: RecentActivity = RecentActivity(text="No episodes yet", time=None)
+
+    @computed_field
+    def total_storage(self) -> str:
+        """Human-readable total storage size derived from total_size (bytes)."""
+        return format_file_size(self.total_size)
+
+
+class PodcastStatistics(BaseModel):
+    """Per-podcast statistics for podcast detail page."""
+
+    episodes_count: int = 0
+    total_duration: int = 0
+    total_size: int = 0
+    last_published_at: datetime | None = None
+    last_created_at: datetime | None = None
